requirements.txt
gdown>=3.10.1
tqdm>=4.30.0
opencv-python>=4.5.5.64
tensorflow>=1.9.0
keras>=2.2.0
flask_cors>=4.0.1
mtcnn>=0.1.0
retina-face>=0.0.1
fire>=0.4.0
streamlit>=1.32.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.18.0
google-generativeai>=0.3.0
numpy>=1.22.0
Pillow>=10.0.0
Flask>=2.3.0
gunicorn>=21.2.0
pymongo>=4.5.0